Updating README + reorganizing the code
